Replace Viper as configuration manager (#12)
Remove Viper from project and add custom config parser.

Viper is not well suited for reading Beacon config - it drops keys without values. It has also other issues and a replacement is needed. Rolling own config parser is questionable and it might need rethinking in the future, but for now the minimal required functionality is provided.